+++ app/controllers/pages_controller.rb
@@ -1,10 +1,11 @@
 class PagesController < ApplicationController
 
-  before_filter { prepend_view_path(File.join(Staytus::Config.theme_root, 'views')) }
+  before_action { prepend_view_path(File.join(Staytus::Config.theme_root, 'views')) }
   layout Staytus::Config.theme_name
 
   def index
-    @services = Service.ordered.includes(:status, {:active_maintenances => :service_status}).to_a
+    @services = Service.ordered.includes(:group, :status, {:active_maintenances => :service_status})
+    @services_with_group = @services.group_by(&:group).sort_by { |g,_| g ? g.name : 'zzz' }
     @issues = Issue.ongoing.ordered.to_a
     @maintenances = Maintenance.open.ordered.to_a
   end
@@ -47,7 +48,7 @@ class PagesController < ApplicationController
     end
   end
 
-  before_filter :check_whether_subscriptions_are_enabled, :only => [:subscribe, :subscribe_by_email]
+  before_action :check_whether_subscriptions_are_enabled, :only => [:subscribe, :subscribe_by_email]
 
   def subscribe
   end